% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-buttonImage.R
\name{buttonImage}
\alias{buttonImage}
\title{Buttons Group Input Control with image}
\usage{
buttonImage(id, labels, values = NULL, active = NULL, file = NULL,
  format = NULL, width = NULL, height = NULL,
  class = "buttonStyle", classImg = "imageStyle")
}
\arguments{
\item{width}{The width of the input, e.g. '400px', or '100\%'.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Input label.}
}
\value{
A buttons group control that can be added to a UI definition.
}
\description{
Create buttons grouped with image.
}
\examples{
\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  codeButton....,
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$somevalue })
}
shinyApp(ui, server)
}
}

}
