% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-imageSelect.R
\name{selectizeInputWithImage}
\alias{selectizeInputWithImage}
\title{Add images to the dropdown options}
\usage{
selectizeInputWithImage(inputId, label, choices, images, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{choices}{List of values to select from.}

\item{images}{List of image location that can be put in a src attribute.}

\item{...}{Arguments that are passed to `selectizeInput()`}
}
\description{
This function will overwrite any render function defined in the options list.
}
