% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyinvoer-package.R
\docType{package}
\name{shinyinvoer-package}
\alias{shinyinvoer}
\alias{shinyinvoer-package}
\title{shinyinvoer: More Input Widgets for Shiny}
\description{
Some more Shiny input widgets built with javascript bindings to provide new input functionalites in Shiny applications, including color palette pickers, image button inputs and predictive search box.
}
\author{
\strong{Maintainer}: Camila Achuri \email{camila@randommonkey.io}

Authors:
\itemize{
  \item David Daza \email{david@randommonkey.io}
  \item Juan Pablo Marín \email{jpmarindiaz@randommonkey.io}
}

}
\keyword{internal}
