% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-coloredSelect.R
\name{updateColoredSelectInput}
\alias{updateColoredSelectInput}
\title{Update Colored Select Input}
\usage{
updateColoredSelectInput(
  session,
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input.}

\item{choices}{The new choices for the input.}

\item{selected}{The new selected value(s).}

\item{placeholder}{The new placeholder text.}
}
\description{
Change the value of a colored select input on the client.
}
