% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-coloredSelectize.R
\name{updateColoredSelectizeInput}
\alias{updateColoredSelectizeInput}
\title{Update Colored Selectize Input}
\usage{
updateColoredSelectizeInput(
  session,
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  colors = NULL,
  placeholder = NULL,
  options = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input.}

\item{choices}{The new choices for the input.}

\item{selected}{The new selected value(s).}

\item{colors}{The new colors for the selected items.}

\item{placeholder}{The new placeholder text.}

\item{options}{The new options for the selectize input.}
}
\description{
Change the value of a colored selectize input on the client.
}
