% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-colorPalette.R
\name{updateColorPaletteInput}
\alias{updateColorPaletteInput}
\title{Update spectrum color picker input}
\usage{
updateColorPaletteInput(session, inputId, colors = list(), palette = list())
}
\description{
Update spectrum color picker input
}
