% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sidePanel.R
\name{updateSidePanelInput}
\alias{updateSidePanelInput}
\title{Update Side Panel Input}
\usage{
updateSidePanelInput(session, inputId, open = NULL, selectedItem = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{open}{Should the panel be open? If \code{NULL}, the state is not
changed.}

\item{selectedItem}{The id of the menu item to select. If \code{NULL}, no
change is made.}
}
\description{
Change the state or content of a side panel input on the client.
}
