% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-spectrumColorPicker.R
\name{updateSpectrumColorPicker}
\alias{updateSpectrumColorPicker}
\title{Update spectrum color picker input}
\usage{
updateSpectrumColorPicker(session, inputId, colors = list(), palette = list())
}
\description{
Update spectrum color picker input
}
