% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-gifLoader.R
\name{withImage}
\alias{withImage}
\title{Add a image that shows when an output is recalculating}
\usage{
withImage(
  ui_element,
  file_location = "https://media.giphy.com/media/vFKqnCdLPNOKc/giphy.gif",
  size = 1
)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a spinner when the corresponding output is being calculated.}

\item{file_location}{The file location of the image file (can be gif or png)}

\item{size}{The size of the gif, relative to its default size.}
}
\description{
Based on [shinycssloaders](https://github.com/daattali/shinycssloaders/blob/master/R/withSpinner.R)
}
\examples{
\dontrun{withImage(plotOutput("my_plot"))}
}
