% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-buttonImage.R
\name{buttonImageInput}
\alias{buttonImageInput}
\title{Buttons Group Input Control with image}
\usage{
buttonImageInput(inputId, labels = NULL, values = NULL,
  active = NULL, file = NULL, format = NULL, class = "buttonStyle",
  classImg = "imageStyle")
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Input label.}

\item{width}{The width of the input, e.g. '400px', or '100\%'.}
}
\value{
A buttons group control that can be added to a UI definition.
}
\description{
Create buttons grouped with image.
}
\examples{
\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  codeButton....,
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$somevalue })
}
shinyApp(ui, server)
}
}

}
