% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-buttonImage.R
\name{buttonImageInput}
\alias{buttonImageInput}
\title{Buttons Group Input Control with image}
\usage{
buttonImageInput(inputId, labels = NULL, values = NULL,
  active = NULL, file = NULL, format = NULL, class = "buttonStyle",
  classImg = "imageStyle")
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{labels}{List of image names saved in a folder into www.}

\item{values}{List of id input for each button.}

\item{active}{Initial button selected.}

\item{file}{Folder where the images that the button will contain are stored.}

\item{class}{CSS class with button css style.}

\item{classImg}{CSS class with image css style.}
}
\value{
A buttons group control that can be added to a UI definition.
}
\description{
Create buttons grouped with image to choose a single button and get your id.
}
\examples{
\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
    uiOutput('button'),
    verbatimTextOutput('input_button')
)
server <- function(input, output) {
 # you must crate a file in www for save images (www/img/...)
 output$button <- renderUI({
                  buttonImageInput(inputId = 'chosen_button',
                  labels = c("cat", "dog", "fox"),
                  values = c("cat", "dog", "fox"),
                  active = 'dog',
                  file = "img/")
                  })
print input id of your click
output$input_button <- renderPrint({
                   input$chosen_button
                  })
}
shinyApp(ui, server)
}
}


}
