% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-toggleSwitch.R
\name{toggleSwitchInput}
\alias{toggleSwitchInput}
\title{Switch state button}
\usage{
toggleSwitchInput(
  inputId,
  label = NULL,
  on_label = "ON",
  off_label = "OFF",
  value = FALSE
)
}
\description{
Switch state button
}
